+++ src/test/resources/org/pgcodekeeper/core/it/diff/pg/modify_not_null_new.sql
@@ -9,6 +9,12 @@ CREATE TABLE public.testtable (
     field8 integer,
     field9 integer NOT NULL,
     field10 integer constraint testtable_field10_not_null NOT NULL,
+    very_very_very_very_very_very_very_very_long_column_name integer CONSTRAINT testtable_very_very_very_very_very_very_very_very_long_not_null NOT NULL,
     CONSTRAINT field7_not_null_test NOT NULL field7 NO INHERIT,
     CONSTRAINT testtable_field8_not_null NOT NULL field8
+);
+
+CREATE TABLE public."👨‍👩‍👧‍👦👨‍👩‍👧‍👦" (
+    "🇷🇺🇷🇺🇷🇺" integer CONSTRAINT "👨‍👩‍👧‍👦👨_🇷🇺🇷🇺🇷🇺_not_null" NOT NULL,
+    "шестьдесятбайт_шестьдесятбайт_ш" integer CONSTRAINT "👨‍👩‍👧‍👦_шестьдесятбай_not_null" NOT NULL
 );