+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+- Fixed default name generation for CONSTRAINT and SEQUENCE objects in PostgreSQL.
+
 ## [13.1.0] - 2025-11-18
 
 ### Changed